Move OAuth2 to separate controller and service
