+++ src/main/java/com/drapson/springauthtutorial/adapters/in/api/AuthController.java
@@ -97,57 +97,4 @@ public class AuthController {
 
         return ResponseEntity.ok().build();
     }
-
-    @PostMapping("/oauth2/link")
-    public ResponseEntity<Void> linkOAuthAccounts(@RequestBody @Valid LinkOAuthAccountRequest linkOAuthAccountRequest,
-                                                  HttpServletResponse response) {
-        Optional<AuthTokens> authTokens = authService.linkNewOAuthAccount(new LinkOAuthAccountDto(
-                linkOAuthAccountRequest.shouldLink(),
-                linkOAuthAccountRequest.provider(),
-                linkOAuthAccountRequest.providerId(),
-                linkOAuthAccountRequest.userId()
-        ));
-
-        if (authTokens.isPresent()) {
-            AuthTokens createdTokens = authTokens.get();
-
-            Cookie refreshTokenCookie = cookieUtil.createRefreshTokenCookie(createdTokens.refreshToken());
-            Cookie accessTokenCookie = cookieUtil.createAccessTokenCookie(createdTokens.accessToken());
-
-            response.addCookie(refreshTokenCookie);
-            response.addCookie(accessTokenCookie);
-
-            return ResponseEntity.ok().build();
-        }
-
-        return ResponseEntity.noContent().build();
-    }
-
-    @PostMapping("/oauth2/code/google")
-    public ResponseEntity<?> handleGoogleCode(@RequestBody OAuthCodeRequest request,
-                                              HttpServletResponse response) {
-        GoogleLoginDTO googleLoginDTO = authService
-                .handleGoogleLogin(new OAuthCodeDto(request.code(), request.codeVerifier()));
-
-        return switch (googleLoginDTO.getLoginType()) {
-            case NEW_USER, EXISTING_USER -> {
-                Cookie refreshTokenCookie = cookieUtil
-                        .createRefreshTokenCookie(googleLoginDTO.getRefreshToken());
-
-                Cookie accessTokenCookie = cookieUtil
-                        .createAccessTokenCookie(googleLoginDTO.getAccessToken());
-
-                response.addCookie(refreshTokenCookie);
-                response.addCookie(accessTokenCookie);
-
-                yield ResponseEntity.ok().build();
-            }
-            case POSSIBLE_LINK -> ResponseEntity.status(HttpStatus.CONFLICT)
-                    .body(new OAuth2LinkResponse(
-                            googleLoginDTO.getProviderName(),
-                            googleLoginDTO.getProviderId(),
-                            googleLoginDTO.getUserId()
-                    ));
-        };
-    }
 }
