+++ src/main/java/com/drapson/springauthtutorial/application/AuthService.java
@@ -18,9 +18,5 @@ public interface AuthService {
 
     boolean checkIfUserHasProvider(User user, String provider);
 
-    Optional<AuthTokens> linkNewOAuthAccount(LinkOAuthAccountDto linkOAuthAccountDto);
-
     String issueTemporaryRegistrationToken(PendingOAuthRegistration pendingOAuthRegistration);
-
-    GoogleLoginDTO handleGoogleLogin(OAuthCodeDto oAuthCodeDto);
 }
